+++ unpythonic/syntax/__init__.py
@@ -30,6 +30,7 @@ from .dbg import (dbg_block as _dbg_block, dbg_expr as _dbg_expr,  # noqa: F401
 from .prefix import prefix as _prefix
 from .tailtools import (autoreturn as _autoreturn, tco as _tco,  # noqa: F401
                         continuations as _continuations, call_cc)
+from .testutil import (test as _test)
 
 # "where" is only for passing through (export).
 from .letdoutil import UnexpandedLetView, _canonize_bindings, where  # noqa: F401
@@ -1972,5 +1973,107 @@ def prefix(tree, **kw):  # noqa: F811
 
 # TODO: using some name other than "kw" would silence the IDE warnings.
 from .prefix import q, u, kw  # for re-export only  # noqa: F401
+
+# -----------------------------------------------------------------------------
+
+# TODO: Some of this description comes from `unpythonic_assert`. We could expose that and refer to it?
+@macros.expr
+def test(tree, **kw):  # noqa: F811
+    """[syntax, expr] Perform a test.
+
+    A low-level tool for building test frameworks for macro-enabled code.
+
+    **Why**:
+
+      `pytest` achieves compact syntax by hijacking `assert`, but macro-enabled
+      code needs MacroPy's macro expander, so we can't allow a testing tool to
+      install its own import hook to do AST transformations (because doing that
+      disables the macro expander).
+
+      And MacroPy macros only exist in expr, block and decorator variants.
+      We can't just hijack any node type of the AST willy-nilly like `pytest`
+      does, so we solve this the MacroPy way - by providing an expr macro
+      that can be used instead of `assert` when writing test cases.
+
+    **Syntax**::
+
+        `test[expr]`
+        `test[expr, "name of my test"]`
+
+    **How to use**:
+
+    The behavior is similar to the builtin `assert`, with a twist.
+
+    If the dynvar `test_signal_errors` is truthy, then, upon a failing
+    test, this will signal the `AssertionError` as a correctable error,
+    via unpythonic's condition system (see `unpythonic.conditions.cerror`).
+
+    If that dynvar is falsey (default), the `AssertionError` is raised.
+
+    The idea is that `cerror` allows surrounding code to install a handler that
+    invokes the `proceed` restart after performing any logging and such, so that
+    although the error is reported, further tests still run::
+
+        from unpythonic.syntax import macros, test
+
+        import sys
+        from unpythonic import dyn, handlers, invoke
+
+        errors = 0
+        def report(err):
+            global errors
+            errors += 1
+            print(err, file=sys.stderr)  # or log or whatever
+            invoke("proceed")
+
+        with dyn.let(test_signal_errors=True):  # use conditions instead of exceptions
+            with handlers((AssertionError, report)):
+                test[2 + 2 == 5]  # fails, but allows further tests to continue
+                test[2 + 2 == 4]
+                test[17 + 23 == 40, "my named test"]
+
+        assert errors == 1
+
+    If you want to proceed after most failures, but there is some particularly
+    critical test which, if it fails, should abort the rest of the whole unit,
+    you can override the handler locally::
+
+        def die(err):
+            print(err, file=sys.stderr)  # or log or whatever
+            sys.exit(255)
+
+        with dyn.let(test_signal_errors=True):  # use conditions instead of exceptions
+            with handlers((AssertionError, report)):
+                test[2 + 2 == 5]  # fails, but allows further tests to continue
+
+                with handlers((AssertionError, die)):
+                    test[2 + 2 == 6]  # --> die
+                    test[17 + 23 == 40, "my named test"]  # not reached
+
+                # if this point was ever reached (currently it's not)...
+                test[2 + 2 == 7]  # ...this fails, but allows further tests to continue
+
+    This works, because the dynamically most recently bound handler for the
+    same signal type wins (see `unpythonic.conditions`).
+
+    Similarly, if you want to skip the rest of a block of tests upon a failure::
+
+        from unpythonic import restarts, invoker
+
+        with dyn.let(test_signal_errors=True):  # use conditions instead of exceptions
+            with handlers((AssertionError, report)):
+                test[2 + 2 == 5]  # fails, but allows further tests to continue
+
+                with restarts(skip=(lambda: None)):  # just for control, no return value
+                    with handlers((AssertionError, invoker("skip"))):
+                        test[2 + 2 == 6]  # --> fails, skip the rest of this block
+                        test[17 + 23 == 40, "my named test"]  # not reached
+
+                test[2 + 2 == 7]  # fails, but allows further tests to continue
+
+    See the unit tests in `unpythonic.syntax.test.test_testutil` for more variations.
+
+    """
+    return _test(tree)
 
 # -----------------------------------------------------------------------------